pyproject.toml
[build-system]
requires = [
  "setuptools>=68",
  "setuptools_scm[toml]>=8.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "smallbiz"
description = "Tools for small business owners"
authors = [{ name = "Sunyoung Yoo" }]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Typing :: Typed",
]
requires-python = ">=3.10"

# IMPORTANT:
# Run 'tox -e deps' after making changes here. This will update requirement files.
# Make sure to list one dependency per line.
dependencies = [
    "pandas",
    "msoffcrypto-tool",
    "openpyxl",
    "xlrd",
    "rich"
]

dynamic = ["version"]

[project.urls]
"Bug Tracker" = "https://github.com/YooSunyoung/smallbiz/issues"
# "Documentation" = "https://YooSunyoung.github.io/smallbiz"
"Source" = "https://github.com/yoosunyoung/smallbiz"

[project.scripts]
merge-orders = "smallbiz_kr.executables.merge_orders:main"

[tool.setuptools_scm]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = """
--strict-config
--strict-markers
--import-mode=importlib
-ra
-v
"""
testpaths = "tests"
filterwarnings = [
  "error",
]

[tool.ruff]
line-length = 88
extend-include = ["*.ipynb"]
extend-exclude = [
    ".*", "__pycache__", "build", "dist", "install",
]

[tool.ruff.lint]
# See https://docs.astral.sh/ruff/rules/
select = ["B", "C4", "DTZ", "E", "F", "G", "I", "PERF", "PGH", "PT", "PYI", "RUF", "S", "T20", "UP", "W"]
ignore = [
    # Conflict with ruff format, see
    # https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "COM812", "COM819", "D206", "D300", "E111", "E114", "E117", "ISC001", "ISC002", "Q000", "Q001", "Q002", "Q003", "W191",
]
fixable = ["I001", "B010"]
isort.known-first-party = ["{{projectname}}"]
pydocstyle.convention = "numpy"

[tool.ruff.lint.per-file-ignores]
# those files have an increased risk of relying on import order
"__init__.py" = ["I"]
"tests/*" = [
    "S101",  # asserts are fine in tests
    "B018",  # 'useless expressions' are ok because some tests just check for exceptions
]
"*.ipynb" = [
    "E501",  # longer lines are sometimes more readable
    "F403",  # *-imports used with domain types
    "F405",  # linter may fail to find names because of *-imports
    "I",  # we don't collect imports at the top
    "S101",  # asserts are used for demonstration and are safe in notebooks
    "T201",  # printing is ok for demonstration purposes
]

[tool.ruff.format]
quote-style = "preserve"

[tool.mypy]
strict = true
ignore_missing_imports = true
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",
]
warn_unreachable = true
